+++ keras_cv/models/__init__.py
@@ -96,30 +96,30 @@ from keras_cv.models.object_detection.retina_net.retina_net import RetinaNet
 from keras_cv.models.object_detection_3d.center_pillar import (
     MultiHeadCenterPillar,
 )
-from keras_cv.models.regnet import RegNetX002
-from keras_cv.models.regnet import RegNetX004
-from keras_cv.models.regnet import RegNetX006
-from keras_cv.models.regnet import RegNetX008
-from keras_cv.models.regnet import RegNetX016
-from keras_cv.models.regnet import RegNetX032
-from keras_cv.models.regnet import RegNetX040
-from keras_cv.models.regnet import RegNetX064
-from keras_cv.models.regnet import RegNetX080
-from keras_cv.models.regnet import RegNetX120
-from keras_cv.models.regnet import RegNetX160
-from keras_cv.models.regnet import RegNetX320
-from keras_cv.models.regnet import RegNetY002
-from keras_cv.models.regnet import RegNetY004
-from keras_cv.models.regnet import RegNetY006
-from keras_cv.models.regnet import RegNetY008
-from keras_cv.models.regnet import RegNetY016
-from keras_cv.models.regnet import RegNetY032
-from keras_cv.models.regnet import RegNetY040
-from keras_cv.models.regnet import RegNetY064
-from keras_cv.models.regnet import RegNetY080
-from keras_cv.models.regnet import RegNetY120
-from keras_cv.models.regnet import RegNetY160
-from keras_cv.models.regnet import RegNetY320
+from keras_cv.models.backbones.regnet import RegNetX002
+from keras_cv.models.backbones.regnet import RegNetX004
+from keras_cv.models.backbones.regnet import RegNetX006
+from keras_cv.models.backbones.regnet import RegNetX008
+from keras_cv.models.backbones.regnet import RegNetX016
+from keras_cv.models.backbones.regnet import RegNetX032
+from keras_cv.models.backbones.regnet import RegNetX040
+from keras_cv.models.backbones.regnet import RegNetX064
+from keras_cv.models.backbones.regnet import RegNetX080
+from keras_cv.models.backbones.regnet import RegNetX120
+from keras_cv.models.backbones.regnet import RegNetX160
+from keras_cv.models.backbones.regnet import RegNetX320
+from keras_cv.models.backbones.regnet import RegNetY002
+from keras_cv.models.backbones.regnet import RegNetY004
+from keras_cv.models.backbones.regnet import RegNetY006
+from keras_cv.models.backbones.regnet import RegNetY008
+from keras_cv.models.backbones.regnet import RegNetY016
+from keras_cv.models.backbones.regnet import RegNetY032
+from keras_cv.models.backbones.regnet import RegNetY040
+from keras_cv.models.backbones.regnet import RegNetY064
+from keras_cv.models.backbones.regnet import RegNetY080
+from keras_cv.models.backbones.regnet import RegNetY120
+from keras_cv.models.backbones.regnet import RegNetY160
+from keras_cv.models.backbones.regnet import RegNetY320
 from keras_cv.models.segmentation.deeplab import DeepLabV3
 from keras_cv.models.stable_diffusion import StableDiffusion
 from keras_cv.models.stable_diffusion import StableDiffusionV2
